--- src/main/java/tutorlink/command/attendancecommand/AddAttendanceCommand.java
@@ -1,18 +1,0 @@
-package tutorlink.command.attendancecommand;
-
-import tutorlink.attendance.Attendance;
-
-import java.util.ArrayList;
-
-public class AddAttendanceCommand extends AttendanceCommand{
-
-    public AddAttendanceCommand(Attendance attendance) {
-        super(attendance);
-    }
-
-    @Override
-    public void execute(ArrayList attendanceList) {
-
-    }
-
-}
