--- src/main/java/tutorlink/command/attendancecommand/AttendanceCommand.java
@@ -1,17 +1,0 @@
-package tutorlink.command.attendancecommand;
-
-import tutorlink.attendance.Attendance;
-import tutorlink.command.Command;
-
-public abstract class AttendanceCommand extends Command {
-    private Attendance attendance;
-
-    public AttendanceCommand(Attendance attendance) {
-        this.attendance = attendance;
-    }
-
-    @Override
-    public boolean isExit() {
-        return false;
-    }
-}
