--- src/main/java/tutorlink/command/attendancecommand/DeleteAttendanceCommand.java
@@ -1,17 +1,0 @@
-package tutorlink.command.attendancecommand;
-
-import tutorlink.attendance.Attendance;
-
-import java.util.ArrayList;
-
-public class DeleteAttendanceCommand extends AttendanceCommand{
-
-    public DeleteAttendanceCommand(Attendance attendance) {
-        super(attendance);
-    }
-
-    @Override
-    public void execute(ArrayList list) {
-
-    }
-}
