--- src/main/java/tutorlink/command/studentcommand/StudentCommand.java
@@ -1,15 +1,0 @@
-package tutorlink.command.studentcommand;
-
-import tutorlink.command.Command;
-import tutorlink.student.Student;
-
-public abstract class StudentCommand extends Command {
-    private Student student;
-    public StudentCommand(Student student) {
-        this.student = student;
-    }
-    @Override
-    public boolean isExit() {
-        return false;
-    }
-}
