--- src/main/java/tutorlink/command/studentcommand/addStudentCommand.java
@@ -1,18 +1,0 @@
-package tutorlink.command.studentcommand;
-
-import tutorlink.student.Student;
-
-import java.util.ArrayList;
-
-public class addStudentCommand extends StudentCommand{
-
-    public addStudentCommand(Student student) {
-        super(student);
-    }
-
-    @Override
-    public void execute(ArrayList list) {
-
-    }
-
-}
