+++ src/main/java/tutorlink/commandpackage/Command.java
@@ -19,3 +19,4 @@ public abstract class Command {
      */
     public abstract boolean isExit();
 }
+
