--- src/main/java/tutorlink/parser/Parser.java
@@ -1,4 +1,0 @@
-package tutorlink.parser;
-
-public class Parser {
-}
