+++ src/test/java/TutorLink/listpackage/AssignmentListTest.java
@@ -1,9 +1,8 @@
-package listpackage;
+package tutorlink.listpackage;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import tutorlink.assignmentpackage.Assignment;
-import tutorlink.listpackage.AssignmentList;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.assertFalse;
