+++ src/test/java/tutorlink/command/assignmentcommand/AddAssignmentCommandTest.java
@@ -1,23 +1,24 @@
 package tutorlink.command.assignmentcommand;
 
 import org.junit.jupiter.api.Test;
-import tutorlink.assignment.Assignment;
+import tutorlink.assignmentpackage.Assignment;
+import tutorlink.commandpackage.assignmentcommand.AddAssignmentCommand;
+import tutorlink.listpackage.AssignmentList;
 
-import java.util.ArrayList;
-
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
 
 class AddAssignmentCommandTest {
     @Test
     void execute_addOne_expectOne() {
         // setup
-        ArrayList<Assignment> assignmentList = new ArrayList<>();
+        AssignmentList assignmentList = new AssignmentList();
         Assignment assignment1 = new Assignment("Test_1",100,100,0.2);
         AddAssignmentCommand addAssignmentCommand = new AddAssignmentCommand(assignment1);
         // execute
         addAssignmentCommand.execute(assignmentList);
-        assertFalse(assignmentList.isEmpty());
-        assertEquals(1, assignmentList.size());
-        assertEquals(assignment1, assignmentList.get(0));
+        assertFalse(assignmentList.getAssignmentArrayList().isEmpty());
+        assertEquals(1, assignmentList.getAssignmentArrayList().size());
+        assertEquals(assignment1, assignmentList.getAssignmentArrayList().get(0));
     }
 }
