--- src/test/java/tutorlink/list/AttendanceListTest.java
@@ -1,4 +1,0 @@
-package tutorlink.list;
-
-public class AttendanceListTest {
-}
