--- src/test/java/tutorlink/list/StudentListTest.java
@@ -1,4 +1,0 @@
-package tutorlink.list;
-
-public class StudentListTest {
-}
